Added my files for the project
trellolist.html -> HTML for webpage with lists and corresponding buttons
trelloliststyle.css -> Style guide for HTML
addnewlist.js -> Instructions for HTML file on adding or removing lists when requested by the user
draganddrop.js -> Instructions for HTML file on dragging and dropping nodes onto other lists

NOTE: there are some bugs with these files but they generally work as they are supposed to. For now, dragging a node onto another node removes it, so be careful and make sure it is being dragged onto the grey area in order to add the node to the list properly.